+++ Diary/Resource/Controller/DetailViewController.swift
@@ -1,0 +1,17 @@
+//
+//  DetailViewController.swift
+//  Diary
+//
+//  Created by SummerCat and som on 2022/12/26.
+//
+
+import UIKit
+
+class DetailViewController: DiaryItemViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+    
+}
