+++ Diary/Resource/Model/Diary+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Diary+CoreDataProperties.swift
 //  
 //
-//  Created by Jiyoung Lee on 2022/12/26.
+//  Created by SummerCat and som on 2022/12/26.
 //
 //
